Background, Boundaries & More
Implemented simulated movement and background looping.
Added player boundaries.
Added the iceberg object and cannonball sprites.
Added a ScoreFunctionalityWait script which makes it so that the score only starts increasing after the tutorial has begun (Score would originally increase when the PROJECT starts, leading to starting the minigame with potentially millions of points if left idle for long enough).
Set the ParticleMasks game object to be inactive (player was getting hidden behind it, object didn't do anything so this should be fine).

Current Issues: The cannon firing locations continue moving when the player hits against a boundary causing the locations to become off-positioned.